move device functions to device.rs
